+++ src/App.tsx
@@ -2,7 +2,7 @@ import GlobalStyles from 'libs/styles/globalStyles';
 import React from 'react';
 import styled from 'styled-components';
 import './App.css';
-import MobileHeader from 'components/MobileHeader';
+import MobileHeader from 'components/base/MobileHeader';
 import Routing from './routes/Routing';
 
 const AsideLayout = styled.div`
@@ -13,9 +13,9 @@ const AsideLayout = styled.div`
 function App() {
   return (
     <div className="App">
+      <GlobalStyles />
       <MobileHeader />
       <AsideLayout>
-        <GlobalStyles />
         <Routing />
       </AsideLayout>
     </div>
