+++ src/components/base/WebHeader.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function WebHeader() {
+  return <div>WebHeader</div>;
+}
+
+export default WebHeader;
